Move code around and separate approval checks
