+++ owners/src/github.js
@@ -15,13 +15,33 @@
  */
 
 const {RepoFile} = require('./repo-file');
+const {OwnersCheck} = require('./owners_check');
 const {Owner} = require('./owner');
 const _ = require('lodash');
 const sleep = require('sleep-promise');
 
 const GITHUB_CHECKRUN_DELAY = 2000;
-const GITHUB_CHECKRUN_NAME = 'ampproject/owners-check';
 const OWNERS_CHECKRUN_REGEX = /owners bot|owners-check/i;
+
+/**
+ * A a GitHub PR review.
+ */
+class Review {
+  /**
+   * Constructor.
+   *
+   * @param {!string} reviewer username of the reviewer giving approval.
+   * @param {!string} state status of the review (ie. "approved" or not).
+   * @param {!Date} submittedAt timestamp when the review was submitted.
+   */
+  constructor(reviewer, state, submittedAt) {
+    Object.assign(this, {
+      reviewer,
+      submittedAt,
+      isApproved: state.toLowerCase() === 'approved',
+    });
+  }
+}
 
 /**
  * Interface for working with the GitHub API.
@@ -150,117 +170,6 @@ class GitHub {
 }
 
 /**
- * Manages checking if a PR has the necessary approvals.
- */
-class ApprovalCheck {
-  /**
-   * Builds a check-run.
-   *
-   * @param {!object} fileOwners ownership rules.
-   * @param {!object} approvers list of usernames that approved this PR.
-   * @return {CheckRun} a check-run based on the approval state.
-   */
-  static buildCheckRun(fileOwners, approvers) {
-    const passing = this._allFilesApproved(fileOwners, approvers);
-    const text = this._buildOutputText(fileOwners, approvers);
-    return new CheckRun(passing, text);
-  }
-
-  /**
-   * Tests if all files are approved by at least one owner.
-   *
-   * @private
-   * @param {!object} fileOwners ownership rules.
-   * @param {!string[]} approvers list of usernames that approved this PR.
-   * @return {boolean} if all files are approved.
-   */
-  static _allFilesApproved(fileOwners, approvers) {
-    return Object.values(fileOwners)
-        .map(fileOwner => fileOwner.owner.dirOwners)
-        .every(dirOwners => !!_.intersection(dirOwners, approvers).length);
-  }
-
-  /**
-   * Build the check-run output comment.
-   *
-   * @private
-   * @param {!object} fileOwners ownership rules.
-   * @param {!string[]} approvers list of usernames that approved this PR.
-   * @return {string} check-run output text.
-   */
-  static _buildOutputText(fileOwners, approvers) {
-    const unapprovedFileOwners = Object.values(fileOwners)
-      .filter(fileOwner =>
-        // Omit sections that has a required reviewer who has approved.
-        !_.intersection(approvers, fileOwner.owner.dirOwners).length)
-
-    const reviewerSuggestions = unapprovedFileOwners
-      .map((fileOwner) => {
-        const reviewers = fileOwner.owner.dirOwners.join(', ');
-        const header = `## possible reviewers: ${reviewers}`;
-        const files = fileOwner.files.map(file => ` - ${file.path}`);
-        return [header, ...files].join('\n');
-      });
-
-    return reviewerSuggestions.join('\n\n');
-  }
-}
-
-/**
- * A GitHub presubmit check-run.
- */
-class CheckRun {
-  /**
-   * Constructor.
-   *
-   * @param {!string} passing whether or not the check-run is passing/approved.
-   * @param {!string} text description of check-run results.
-   */
-  constructor(passing, text) {
-    Object.assign(this, {passing, text});
-  }
-
-  /**
-   * Produces a JSON version of the object for use with GitHub API.
-   *
-   * @return {object} JSON object describing check-run.
-   */
-  get json() {
-    return {
-      name: GITHUB_CHECKRUN_NAME,
-      status: 'completed',
-      conclusion: 'neutral',
-      completed_at: new Date(),
-      output: {
-        title: GITHUB_CHECKRUN_NAME,
-        summary: `The check was a ${this.passing ? 'success' : 'failure'}!`,
-        text: this.text,
-      },
-    };
-  }
-}
-
-/**
- * A a GitHub PR review.
- */
-class Review {
-  /**
-   * Constructor.
-   *
-   * @param {!string} reviewer username of the reviewer giving approval.
-   * @param {!string} state status of the review (ie. "approved" or not).
-   * @param {!Date} submittedAt timestamp when the review was submitted.
-   */
-  constructor(reviewer, state, submittedAt) {
-    Object.assign(this, {
-      reviewer,
-      submittedAt,
-      isApproved: state.toLowerCase() === 'approved',
-    });
-  }
-}
-
-/**
  * Maps the github json payload to a simpler data structure.
  */
 class PullRequest {
@@ -292,7 +201,7 @@ class PullRequest {
     const approvers = await this.getApprovers();
 
     const checkRunId = await this.github.getCheckRunId(this.headSha);
-    const latestCheckRun = ApprovalCheck.buildCheckRun(fileOwners, approvers);
+    const latestCheckRun = OwnersCheck.buildCheckRun(fileOwners, approvers);
 
     if (checkRunId) {
       await this.github.updateCheckRun(checkRunId, latestCheckRun);
@@ -360,9 +269,4 @@ class PullRequest {
   }
 }
 
-module.exports = {
-  CheckRun,
-  GitHub,
-  PullRequest,
-  Review,
-};
+module.exports = {GitHub, PullRequest, Review};
